Update service logic according code review suggestions
